create tracking components and styles
. made components and distinguish their styles
. set global value in one place
