+++ src/components/utils/Globals.js
@@ -1,0 +1,18 @@
+// @flow
+
+module.exports = {
+    COLOR: {
+        WHITE: '#FFFFFF',
+        GREEN: 'green',
+        RED: '#FF0000',
+        PRIMARY: '#0F303E'
+    },
+    STRINGS: {
+        TRACK_SHIPMENT: 'Track Shipment',
+    },
+    SIZE: {
+        DEFAULT_MARGIN: 16,
+        STANDARD_PADDING: 16,
+        TOOLBAR_HT: 56,
+    }
+};
